Update AUTHORS.md
Hello!  Just a few corrections around the original ownership.  I (Eric) am actually the originally author of devmon, having written it while working as a Network Engineer for the University of Virginia medical center back in 2005.

I handed off the reins to Buchan (whom I'm assuming is the one that you contacted) some time in 2007 when I left the University of Virginia (and subsequently had nothing left to monitor via SNMP).    You can probably confirm all of this by going back and looking through the Sourceforge support group, if you _really_ want to.

At any rate, it's lovely to see that somebody is still using it, and maintaining it!  You have my blessing to take over ownership =D